refactor(probe): migrate ndp planning into per-protocol plugin
